pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-metadata-framework"
version = "0.0.1"
description = "Metadata-driven runtime extension for dagster-dag-factory"
dependencies = [
    "dagster==1.11.1",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    # We expect dagster-dag-factory to be installed as an editable or in path
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
]
